+++ CHANGELOG.md
@@ -11,10 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - French localisation
 - Support for PostgreSQL ([#503], [#504])
 - Warning for waiting room settings ([#302], [#509])
+- Option to disable local authentication ([#505], [#506])
 
 ### Changed
 - Bump dependencies
 - Improve performance on permission checks ([#469])
+- Renamed env option PASSWORD_SELF_RESET_ENABLED to PASSWORD_CHANGE_ALLOWED and enabled by default (PASSWORD_SELF_RESET_ENABLED deprecated) ([#506])
 
 ### Fixed
 - Network error on logout due to reload of all components
@@ -629,6 +631,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#470]: https://github.com/THM-Health/PILOS/pull/470
 [#503]: https://github.com/THM-Health/PILOS/issues/503
 [#504]: https://github.com/THM-Health/PILOS/pull/504
+[#505]: https://github.com/THM-Health/PILOS/issues/505
+[#506]: https://github.com/THM-Health/PILOS/pull/506
 [#509]: https://github.com/THM-Health/PILOS/pull/509
 
 [unreleased]: https://github.com/THM-Health/PILOS/compare/v2.2.0...HEAD